Did some resizing and bug fixing
I resized the sidebar yet again to make more room for hhe Announcement and Trending sections. I also managed to get a single-column layout when zooming in all the way on the project sections. I had an issue where row gaps disappeared in the projects section as I zoomed in but I added overflow:hidden on the Project_Content div to take care of that. Next, I will work on the Announcements and Trending sections.